move styles to assets
