+++ client/src/assets/styles/shared-styles.ts
