+++ client/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './styles/index.css';
+import './assets/styles/index.css';
 import App from './App';
 import {
   Route,
